+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/Util/Util.java
@@ -1,0 +1,15 @@
+package com.rperazzo.weatherapp.Util;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+
+public class Util {
+
+    public static boolean isDeviceConnected(Context c) {
+        ConnectivityManager cm = (ConnectivityManager)
+                c.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo netInfo = cm.getActiveNetworkInfo();
+        return netInfo != null && netInfo.isConnected();
+    }
+}
